+++ src/main/java/com/universestay/project/admin/dao/NoticeDao.java
@@ -1,6 +1,6 @@
-package com.universestay.project.notice.dao;
+package com.universestay.project.admin.dao;
 
-import com.universestay.project.notice.dto.NoticeDto;
+import com.universestay.project.admin.dto.NoticeDto;
 import java.util.List;
 
 public interface NoticeDao {
