+++ src/main/java/com/universestay/project/admin/dto/NoticeDto.java
@@ -1,4 +1,4 @@
-package com.universestay.project.notice.dto;
+package com.universestay.project.admin.dto;
 
 
 import io.swagger.annotations.ApiModel;
