+++ src/main/resources/mybatis/mapper/admin/noticeMapper.xml
@@ -2,14 +2,14 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.universestay.project.notice.dao.NoticeDao">
-    <select id="selectAll" resultType="com.universestay.project.notice.dto.NoticeDto">
+<mapper namespace="com.universestay.project.admin.dao.NoticeDao">
+    <select id="selectAll" resultType="NoticeDto">
         select *
         from Notice
         order by notice_id DESC
     </select>
 
-    <select id="select" resultType="com.universestay.project.notice.dto.NoticeDto">
+    <select id="select" resultType="NoticeDto">
         select *
         from Notice
         where notice_id = #{notice_id}
@@ -21,13 +21,13 @@
         where notice_id = #{notice_id}
     </delete>
 
-    <insert id="insert" parameterType="com.universestay.project.notice.dto.NoticeDto">
+    <insert id="insert" parameterType="NoticeDto">
         insert into Notice
         (admin_id, notice_title, notice_ctt, created_at, created_id, updated_at, updated_id)
         values (#{admin_id}, #{notice_title}, #{notice_ctt}, now(), #{admin_id}, now(), #{admin_id})
     </insert>
 
-    <update id="update" parameterType="com.universestay.project.notice.dto.NoticeDto">
+    <update id="update" parameterType="NoticeDto">
         update Notice
         set notice_title   = #{notice_title},
             notice_ctt     = #{notice_ctt},
