+++ src/main/resources/mybatis/mybatis-config.xml
@@ -7,7 +7,13 @@
         <typeAlias alias="UserDto" type="com.universestay.project.user.dto.UserDto"/>
         <typeAlias alias="SearchCondition"
                 type="com.universestay.project.admin.dto.SearchCondition"/>
+
+        
         <typeAlias alias="AdminDto" type="com.universestay.project.admin.dto.AdminDto"/>
         <typeAlias alias="EventDto" type="com.universestay.project.admin.dto.EventDto"/>
+        <typeAlias alias="CouponDto" type="com.universestay.project.admin.dto.CouponDto"/>
+        <typeAlias alias="InquiryDto" type="com.universestay.project.admin.dto.InquiryDto"/>
+        <typeAlias alias="NoticeDto" type="com.universestay.project.admin.dto.NoticeDto"/>
+
     </typeAliases>
 </configuration>
